+++ astro.config.mjs
@@ -1,13 +1,14 @@
-import { defineConfig } from 'astro/config'
-import sitemap from '@astrojs/sitemap'
-import tailwind from '@astrojs/tailwind'
+import { defineConfig } from 'astro/config';
+import sitemap from '@astrojs/sitemap';
+import tailwind from '@astrojs/tailwind';
+import vercel from '@astrojs/vercel/static';
 
-import vercel from '@astrojs/vercel/static'
+import react from "@astrojs/react";
 
 // https://astro.build/config
 export default defineConfig({
   site: 'https://elfrijolmagico.cl',
-  integrations: [sitemap(), tailwind()],
+  integrations: [sitemap(), tailwind(), react()],
   output: 'static',
   adapter: vercel()
-})
+});
