+++ package.json
@@ -10,14 +10,19 @@
     "astro": "astro"
   },
   "dependencies": {
-    "@astrojs/sitemap": "^3.0.3",
-    "@astrojs/tailwind": "^5.0.3",
-    "@astrojs/vercel": "^6.1.0",
-    "astro": "^4.0.6",
+    "@astrojs/react": "^3.0.9",
+    "@astrojs/sitemap": "^3.0.4",
+    "@astrojs/tailwind": "^5.1.0",
+    "@astrojs/vercel": "^6.1.2",
+    "@types/react": "^18.2.46",
+    "@types/react-dom": "^18.2.18",
+    "astro": "^4.0.9",
     "normalize.css": "^8.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "sharp": "^0.33.1",
     "simplycountdown.js": "^1.7.0",
     "swiper": "^11.0.5",
-    "tailwindcss": "^3.3.7"
+    "tailwindcss": "^3.4.0"
   }
 }
