+++ src/components/H1.astro
@@ -1,0 +1,15 @@
+<h1
+  class="text-6xl md:text-9xl xl:text-[12rem] 2xl:text-[16rem] block font-thickedy tracking-tight text-transparent bg-clip-text bg-[url(/m5.webp)] bg-top font-black [text-shadow:0_0_.4rem_#ffcd6950] sm:[text-shadow:0_0_2rem_#ffcd6950]"
+>
+  <slot />
+</h1>
+<style>
+  h1 {
+    animation: paper 5s steps(5) infinite;
+  }
+  @keyframes paper {
+    to {
+      background-position: bottom;
+    }
+  }
+</style>
