+++ src/components/NavButton.astro
@@ -1,0 +1,30 @@
+<input type="checkbox" id="toggle-nav-btn" hidden class="peer" />
+<label
+  for="toggle-nav-btn"
+  class="relative peer-checked:fixed w-fit outline-none z-50 peer-checked:[&>.first]:hidden peer-checked:[&>.last]:block cursor-pointer transition hover:scale-110 peer-checked:-rotate-90 rotate-0 block xl:hidden"
+>
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+    stroke="currentColor"
+    stroke-width="2"
+    class="w-8 h-8 first"
+    viewBox="0 0 24 24"
+    ><path
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      d="M4 6h16M4 12h8m-8 6h16"></path></svg
+  >
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    fill="none"
+    stroke="currentColor"
+    stroke-width="2"
+    class="hidden w-8 h-8 last"
+    viewBox="0 0 24 24"
+    ><path
+      stroke-linecap="round"
+      stroke-linejoin="round"
+      d="M6 18 18 6M6 6l12 12"></path></svg
+  >
+</label>
