+++ src/components/PodcastCard.astro
@@ -1,10 +1,13 @@
 ---
-const { podcastId } = Astro.props;
+interface Props {
+  id: string;
+}
+const { id } = Astro.props;
 ---
 
 <iframe
   class="w-full border-none h-20 sm:h-[232px]"
   title="Podcasts Frijol Mágico"
-  src={`https://open.spotify.com/embed/episode/${podcastId}?utm_source=generator&t=0`}
+  src={`https://open.spotify.com/embed/episode/${id}?utm_source=generator&t=0`}
   allow="autoplay; clipboard-write; encrypted-media; fullscreen; picture-in-picture"
   loading="lazy"></iframe>
