+++ src/components/Section.astro
@@ -14,9 +14,11 @@ const lowerHref = href ? href : `/${name.toLowerCase()}`;
 ---
 
 <section class="h-fit w-full grid gap-4 max-w-5xl mx-auto">
-  <div class="flex justify-between items-end px-4 sm:px-8 gap-2 w-full">
+  <div
+    class="flex flex-col sm:flex-row justify-between sm:items-end px-4 sm:px-8 gap-4 w-full"
+  >
     <h2
-      class="text-2xl sm:text-4xl lg:text-5xl leading-none font-galindo font-normal tracking-normal max-w-xl -mb-2"
+      class="text-2xl md:text-5xl lg:text-5xl leading-none font-galindo font-normal tracking-normal max-w-xl -mb-2"
     >
       {title}
     </h2>
