+++ src/components/Slider.astro
@@ -8,9 +8,9 @@ import Link from "./Link.astro";
   class="swiper w-full h-full touch-pan-y mx-auto space-y-8 relative overflow-hidden block"
 >
   <header
-    class="flex items-end justify-between font-galindo px-4 sm:px-8 gap-2 w-full max-w-5xl mx-auto"
+    class="flex flex-col sm:flex-row sm:items-end justify-between font-galindo px-4 sm:px-8 gap-4 w-full max-w-5xl mx-auto"
   >
-    <h2 class="text-2xl sm:text-5xl leading-none">Conoce a lxs Ilustradoxs</h2>
+    <h2 class="text-2xl md:text-5xl leading-none">Conoce a lxs Artistas</h2>
     <Link text={`Ir a Ilustradorxs`} href="/ilustradorxs" />
   </header>
   <ul class="swiper-wrapper relative w-full h-full flex transition-transform">
