+++ src/components/SliderCard.astro
@@ -15,7 +15,7 @@ const firstImage = getImagesByName(artist)[0];
 ---
 
 <li
-  class="swiper-slide shrink-0 w-fit h-full cursor-grab px-10 lg:px-32 opacity-20 transition-opacity duration-500"
+  class="swiper-slide shrink-0 w-fit h-full cursor-grab px-4 md:px-12 lg:px-32 opacity-20 transition-opacity duration-500"
 >
   <div
     class:list={[
@@ -32,7 +32,7 @@ const firstImage = getImagesByName(artist)[0];
     <Link
       text={artist}
       href={"/illustradorxs/" + id}
-      class="absolute bottom-5 left-5"
+      class="absolute bottom-3 left-3 sm:bottom-5 sm:left-5"
     />
   </div>
 </li>
