+++ src/components/Title.astro
@@ -6,12 +6,12 @@ const { size = "base" } = Astro.props;
 
 const sizes = {
   base: {
-    h1: "pb-20 text-xl sm:text-2xl lg:text-xl",
-    span: "text-[3.5rem] sm:text-8xl lg:text-[16rem]",
+    h1: "pb-20 text-xl md:text-2xl lg:text-xl",
+    span: "text-6xl md:text-9xl xl:text-[12rem] 2xl:text-[16rem] -mt-4 md:-mt-6",
   },
   sm: {
     h1: "pb-20 text-sm sm:text-lg",
-    span: "text-[3.5rem] sm:text-8xl lg:text-[12rem] ",
+    span: "text-5xl md:text-8xl xl:text-[10rem] 2xl:text-[12rem] -mt-4 md:-mt-6",
   },
 };
 ---
@@ -22,7 +22,7 @@ const sizes = {
   Asociación Cultural
   <span
     class:list={[
-      "sm:-mt-6 pb-4 pt-3 block font-thickedy tracking-tight text-transparent bg-clip-text bg-[url(/m5.png)] bg-top font-black [text-shadow:0_0_.4rem_#ffcd6950] sm:[text-shadow:0_0_2rem_#ffcd6950]",
+      "pb-4 pt-3 block font-thickedy tracking-tight text-transparent bg-clip-text bg-[url(/m5.webp)] bg-top font-black [text-shadow:0_0_.4rem_#ffcd6950] sm:[text-shadow:0_0_2rem_#ffcd6950]",
       sizes[size].span,
     ]}>Frijol Mágico</span
   >
