+++ src/data/LIVE_INFO.ts
@@ -1,0 +1,5 @@
+export const LIVE_INFO = [
+  { id: 'dia-1', day: 'Viernes 16', video: 'K0wNji7fWt8' },
+  { id: 'dia-2', day: 'Sábado 17', video: '9IrEYZS9PUo' },
+  { id: 'dia-3', day: 'Domingo 18', video: 'WKRSuM85YEQ' }
+]
