+++ src/data/PODCAST_INFO.ts
@@ -1,0 +1,22 @@
+export const PODCAST_INFO = [
+  {
+    title: 'Como el Pokémon pero con i',
+    id: '2gA0NZpDBSnOK0l79CFzAo',
+    artist: 'Ivichu.jpg'
+  },
+  {
+    title: 'Me salio costureque',
+    id: '7goBEMlWwB0k5wkxrOGeRI',
+    artist: 'Taller de Carroña'
+  },
+  {
+    title: 'Siempre es un gusto de hablar de uno mismo',
+    id: '0boLCWlwKaHG3S2TpV3LEm',
+    artist: 'Rotten Monkey'
+  },
+  {
+    title: 'Comunidad y post pandemia',
+    id: '5wdaKjGEDZRZMUDhbkeDUh',
+    artist: 'Aurora Ravello'
+  }
+]
