+++ src/data/SITE_INFO.ts
@@ -1,0 +1,6 @@
+export const SITE_INFO = {
+  title: 'Apoyando la Ilustración de la Región de Coquimbo!',
+  description:
+    'Frijol Mágico es un espacio que reúne a las y los Ilustradores de la Región de Coquimbo, generando distintas instancias que ayuden a potenciar su trabajo.',
+  logo: '/logo.png'
+}
