--- src/data/podcasts.data.js
@@ -1,18 +1,0 @@
-export const podcastData = [
-  {
-    title: 'Como el Pokémon pero con i',
-    id: '2gA0NZpDBSnOK0l79CFzAo',
-  },
-  {
-    title: 'Me salio costureque',
-    id: '7goBEMlWwB0k5wkxrOGeRI',
-  },
-  {
-    title: 'Siempre es un gusto de hablar de uno mismo',
-    id: '0boLCWlwKaHG3S2TpV3LEm',
-  },
-  {
-    title: 'Comunidad y post pandemia',
-    id: '5wdaKjGEDZRZMUDhbkeDUh',
-  }
-]
