+++ src/layouts/Layout.astro
@@ -39,8 +39,10 @@ const { title = SITE_INFO.title, description = SITE_INFO.description } =
   <body
     class:list={[
       "font-montserrat bg-p-blue relative h-fit w-full min-h-[100dvh] text-white",
-      " after:content-[url(/rama.png)] after:-scale-x-100 after:fixed after:h-full after:-right-1/4 after:bottom-80 after:-z-10",
-      " before:content-[url(/rama.png)] before:fixed before:h-full before:-left-1/4 before:bottom-80 before:-z-10",
+      "",
+      "before:content-[''] sm:before:content-[url(/rama.png)] before:fixed before:h-full sm:before:-left-2/3 lg:before:-left-1/3 xl:before:-left-1/4 before:bottom-80 before:-z-20 lg:before:-z-10",
+      "",
+      "after:content-[url(/rama.png)] after:fixed after:h-full after:-right-[28rem] sm:after:-right-2/3 lg:after:-right-1/3 xl:after:-right-1/4 after:-top-80 after:-z-30 lg:after:-z-20 after:-scale-x-100",
     ]}
   >
     <Header />
@@ -57,16 +59,17 @@ const { title = SITE_INFO.title, description = SITE_INFO.description } =
 
   @font-face {
     font-family: Thickedy;
-    src: url(/fonts/ThickedyQuick.woff2);
+    font-style: normal;
+    font-display: swap;
+    src: url(/fonts/ThickedyQuick.woff2) format("woff2");
   }
 
   body {
-    background-image: url(background.png);
+    background-image: url(/background.webp);
     background-position: top center;
     background-repeat: no-repeat;
-    background-size: 100%;
+    background-size: cover;
     background-attachment: fixed;
-
     background-blend-mode: color;
   }
 </style>
