+++ src/pages/404.astro
@@ -2,7 +2,7 @@
 import Layout from "@layouts/Layout.astro";
 ---
 
-<Layout title="Frijol Mágico | Error 404">
+<Layout title="Error 404">
   <main>
     <h2>
       QUÉ BUSCABAS REALMENTE? ACÁ NO HAY NADA QUE VER, LE SOLICITO QUE VUELVA EN
