+++ src/pages/festivales/index.astro
@@ -1,0 +1,10 @@
+---
+import H1 from "@components/H1.astro";
+import Layout from "@layouts/Layout.astro";
+---
+
+<Layout title="Festivales">
+  <main class="h-screen grid justify-center w-full">
+    <H1>Festivales</H1>
+  </main>
+</Layout>
