--- src/pages/galeria/[artists].astro
@@ -1,31 +1,0 @@
----
-import Layout from "@layouts/Layout.astro";
-import { ARTISTS_INFO } from "@data/ARTISTS_INFO";
-
-export function getStaticPaths() {
-  return ARTISTS_INFO.map((artistInfo) => ({
-    params: { artists: artistInfo.id },
-    props: { ...artistInfo },
-  }));
-}
-
-const { artist, id, contact, images, location, bio } = Astro.props;
----
-
-<Layout title={`Frijol Mágico | ${artist}`}>
-  <main>
-    <h2>En Construcción</h2>
-  </main>
-</Layout>
-
-<style>
-  main {
-    font-size: 1.25rem;
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
-  }
-</style>
